+++ assets/src/scripts/components/hydrograph/time-series-data.js
@@ -54,7 +54,10 @@ export const drawDataLine = function(group, {visible, lines, tsKey, xScale, ySca
             const maskCode = line.classes.dataMask.toLowerCase();
             const maskDisplayName = MASK_DESC[maskCode].replace(' ', '-').toLowerCase();
             const [xDomainStart, xDomainEnd] = extent(line.points, d => d.dateTime);
-            const [yRangeStart, yRangeEnd] = yScale.domain();
+            const [yRangeStart, yRangeEnd] = yScale.range();
+
+            // Some data is shown with the yAxis decreasing from top top bottom
+            const yTop = yRangeEnd > yRangeStart ? yRangeStart : yRangeEnd;
             let maskGroup = group.append('g')
                 .attr('class', `${tsKey}-mask-group`);
             const xSpan = xScale(xDomainEnd) - xScale(xDomainStart);
@@ -62,9 +65,9 @@ export const drawDataLine = function(group, {visible, lines, tsKey, xScale, ySca
 
             maskGroup.append('rect')
                 .attr('x', xScale(xDomainStart))
-                .attr('y', yScale(yRangeEnd))
+                .attr('y', yTop)
                 .attr('width', rectWidth)
-                .attr('height', Math.abs(yScale(yRangeEnd) - yScale(yRangeStart)))
+                .attr('height', Math.abs(yRangeEnd - yRangeStart))
                 .attr('class', `mask ${maskDisplayName}-mask`)
                 .classed(tsKeyClass, true);
 
@@ -73,9 +76,9 @@ export const drawDataLine = function(group, {visible, lines, tsKey, xScale, ySca
 
             maskGroup.append('rect')
                 .attr('x', xScale(xDomainStart))
-                .attr('y', yScale(yRangeEnd))
+                .attr('y', yTop)
                 .attr('width', rectWidth)
-                .attr('height', Math.abs(yScale(yRangeEnd) - yScale(yRangeStart)))
+                .attr('height', Math.abs(yRangeEnd - yRangeStart))
                 .attr('fill', patternId);
         }
     }